requirements.txt
discord.py
aiohttp
isodate
python-dotenv
PyNaCl
cachetools
gtts